+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -23,11 +23,11 @@ export class CartListComponent implements OnInit, DoCheck {
 
   ngOnInit(): void {
     this.cartItems = this.cartService.getProducts$();
+    this.updateData();
   }
 
   ngDoCheck(): void {
-    this.itemsInCart = this.cartService.totalQuantity;
-    this.totalSum = this.cartService.totalSum;
+    this.updateData();
   }
 
   onIncreaseQuantity(id: number): void {
@@ -40,5 +40,10 @@ export class CartListComponent implements OnInit, DoCheck {
 
   onRemoveItem(id: number): void {
     this.cartService.removeProduct(id);
+  }
+
+  private updateData(): void {
+      this.itemsInCart = this.cartService.totalQuantity;
+      this.totalSum = this.cartService.totalSum;
   }
 }