+++ src/app/core/services/local-storage.service.ts
@@ -1,11 +1,16 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+    providedIn: 'root',
+})
 export class LocalStorageService {
     constructor() {}
 
-    getItem(key): string {
-        return window.localStorage.getItem(key);
+    getItem(key): object {
+        return JSON.parse(window.localStorage.getItem(key));
     }
 
     setItem(key: string, value): void {
-        window.localStorage.setItem(key, value);
+        window.localStorage.setItem(key, JSON.stringify(value));
     }
 }