+++ src/app/shared/components/index.ts
@@ -1,0 +1,3 @@
+export * from './first-component/first.component';
+export * from './page-not-found/page-not-found.component';
+export * from './page-forbidden/page-forbidden.component';