+++ src/app/shared/guards/is-cart-empty.guard.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { CartService } from '../../cart/services';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class IsCartEmptyGuard implements CanActivate {
+  constructor(private cartService: CartService) {}
+
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean  {
+    return this.cartService.isEmptyCart();
+  }
+}